qducloud/src/main/java/org/qducloud/bean/Account.java
package org.qducloud.bean;


/**
 * @author Cuiql
 */
//@EntityScan
public class Account {
	private String id;
	private String userName;
	private String password;
	private String uid;
	private String rid;
	public String getId() {
		return id;
	}
	public void setId(String id) {
		this.id = id;
	}
	public String getUserName() {
		return userName;
	}
	public void setUserName(String userName) {
		this.userName = userName;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	public String getUid() {
		return uid;
	}
	public void setUid(String uid) {
		this.uid = uid;
	}
	public String getRid() {
		return rid;
	}
	public void setRid(String rid) {
		this.rid = rid;
	}
	
}
